Extracted rule editor outside of the modal and into its own component
